Feature - Added a scrolling/decaying noise line shader

+++ Source/ComputeShaderExample/Public/Mandelbrot/MandelbrotShaderComponent.h
@@ -66,5 +66,5 @@ public:
 	int Mode = 0;
 
 private:
-	float ZoomLevel;
+	double ZoomLevel;
 };
